Add PCAP_ERROR_CAPTURE_NOTSUP error.
This means "the platform you're on lacks a capture mechanism that
supports that device".

This was inspired by some Stack Overflow questions in which the user was
probably trying to run a capture program on Windows Services for Linux
1, in which the NT kernel attempts to simulate Linux system calls but
does not support packet sockets.

It can, and does, also handle a system in which the BPF capture
mechanism is the native capture mechanism but that has no BPF devices.

The hope is that the error message they get will better indicate the
underlying problem.

(cherry picked from commit 00c52ab6b704620af7ba3663f5af3b89dc0a51fc)