Appveyor: use "choco install" rather than "cinst".
Chocolatey 2.0 got rid of "cinst" and some other wrappers.  It was a
shortcut for "choco install" that was deprecated in 1.0 (but none of the
AppVeyor Windows VMs had 1.x pre-installed, so we didn't see a
deprecation warning that might have indicated that we should be using
"choco install") and whose replacement doesn't appear to be documented
anywhere on the Chocolatey site.

(cherry picked from commit 09230c1db65fcc5c274e4de3e42c7fb7bc1051e6)