Parameterize the interface name in reactivatetest. [skip appveyor]
> uname -sr
Haiku 1
> ./testprogs/reactivatetest
reactivatetest: Neither lo0 nor lo could be opened: lo: No such device
  exists (SIOCGIFSTATS: Invalid Argument)

The loopback interface is "lo" in Linux and "lo0" in BSD systems, but in
Haiku it is "loop", so instead of hard-coding or guessing let's just
make it a mandatory command-line argument.  Also the user may want to
run the test on a non-loopback interface.