Makefile.in: Include instrument-functions.c in the release tarball
Add it in the EXTRA_DIST macro.

Even if it's not needed to configure with '--enable-instrument-functions'
and build, this change helps to use a method similar to that documented
in the FAQ, with a distribution archive.
(FAQ Entry: "I have a program that uses libpcap, is there a way to print
 out debug statements at runtime?")

(backported from commit 239a083dd0be1f0835906e44e5c71ac3b6b597f4)

[skip ci]