sockutils: don't keep our own socket count.
WSAStartup() and WSACleanup() maintain their count of the number of
times WSAStartup() has been called minus the number of times
WSACleanup() has been called.  If WSAStartup() is called when that count
is zero, it initializes Winsock and, whether the count is zero or not,
it increments the count.  If WSACleanup() is called when that count is
1, it shuts down Winsock and whether the count is 1 or not, it
decrements the count.