Improve some error reporting in pcap-snf.c. [skip ci]
The myri_snf kernel module is not loaded, so the host has no snfN
interfaces.  Whether caused by this or not, snf_init() in this case both
sets errno to ENOENT and returns it.  Arguably, this ought to be ENODEV,
but in any case it is more useful to see at least some error message
than just "failed".

Before:
Error in pcap_findalldevs: snf_getifaddrs: snf_init failed

After:
Error in pcap_findalldevs: snf_init: No such file or directory