DAG: In dag_findalldevs() handle known errors better.
pcap_findalldevs() is meant to return all capture devices that libpcap
has detected on the host.  Whether the user has the permissions to use a
particular device is a concern of pcap_activate(), which is equipped to
indicate insufficient permissions as a reason for failure.  The problem
is, dag_findalldevs() does not always conform to this semantics.

One reason for this is that dag_open() requires a write permission on
/dev/dag*.  If the user does not have the permission, all DAG capture
devices are just silently missing from the list:

1.nflog (Linux netfilter log (NFLOG) interface) [none]
2.nfqueue (Linux netfilter queue (NFQUEUE) interface) [none]

To fix this, read the list of card-specific streams from sysfs when
dag_open() returns EACCES.  The result lacks proper descriptions, but is
better than nothing:

1.dag0 [none]
2.dag0:0 [none]
3.dag0:2 [none]
4.nflog (Linux netfilter log (NFLOG) interface) [none]
5.nfqueue (Linux netfilter queue (NFQUEUE) interface) [none]

Another reason for this is that even when the current user has the
permissions (e.g. root) and dag_open() succeeds, dag_attach_stream64()
fails with EBUSY if another process has already attached the same
stream.  Then the stream is missing from the resulting list, and for
stream 0 this makes even less sense because "dag0" means the same as
"dag0:0":

1.dag0 (DAG 7.5G2) [none]
2.dag0:2 (DAG 7.5G2) [none]
3.nflog (Linux netfilter log (NFLOG) interface) [none]
4.nfqueue (Linux netfilter queue (NFQUEUE) interface) [none]

To fix this, treat this particular error as a proof that the stream is a
valid potential capture device and register it normally:

1.dag0 (DAG 7.5G2) [none]
2.dag0:0 (DAG 7.5G2) [none]
3.dag0:2 (DAG 7.5G2) [none]
4.nflog (Linux netfilter log (NFLOG) interface) [none]
5.nfqueue (Linux netfilter queue (NFQUEUE) interface) [none]