DAG: Make device descriptions more useful, make dagN conditional. [skip ci]
In each capture device description make it clear that dagN is a
shorthand for dagN:0 and dagN:M is a receive-only stream M.  When the
user can query /dev/dagN directly, include the card name, the bus
address and the buffer size in each stream description.  Include dagN in
the list iff dagN:0 is included.

Given a host with one physical and two virtual DAG cards, where the
physical card has only stream 2 enabled and one of the virtual cards has
no streams enabled, it looks as follows.

Before (as user):
1.dag0 [none]
2.dag0:2 [none]
3.dag16 [none]
4.dag17 [none]
5.dag17:0 [none]

Before (as root):
1.dag0 (DAG 7.5G2) [none]
2.dag0:2 (DAG 7.5G2) [none]
3.dag16 (vDAG) [none]
4.dag17 (vDAG) [none]
5.dag17:0 (vDAG) [none]

After (as user):
1.dag0:2 (Rx stream 2) [none]
2.dag17 (alias for dag17:0) [none]
3.dag17:0 (Rx stream 0) [none]

After (as root):
1.dag0:2 (Rx stream 2, size: 128 MiB, bus: 0000:01:00.0, name: DAG
  7.5G2) [none]
2.dag17 (alias for dag17:0) [none]
3.dag17:0 (Rx stream 0, size: 64 MiB, bus: VDAG, name: vDAG) [none]