DAG: Remove the strrchr() call from pcap_create(). [skip ci]
The original intention was likely to skip the optional "/dev/" in
"/dev/dag0" to support a larger subset of what dag_parse_name()
implements.  It does not work this way because the pointer strrchr()
returns points to the last '/' and the test for the first three
characters being "dag" in "/dag0" is always negative:

$ tcpdump -ni /dev/dag0
tcpdump: This version of libpcap only supports DAG cards

Remove the broken code for now to simplify the module refactoring.  If
there is a good reason to support DAG device aliases in addition to the
nominal names in libpcap, it will not be difficult to add that.