Remove remaining no-op strrchr() calls.
Apparently, its intended purpose indeed is to skip an optional "/dev/"
before the capture device name, but only in open_dlpi_device(), which
seems to handle the return value correctly.  The instances in other
modules are no-op side effects of commit 2426611.  Eliminate the local
variable as well where it makes sense.