DAG: Refine error handling in dag_findalldevs().
The only purpose of the dag_parse_name() invocation in this context is
to derive "/dev/dagN" from "dagN", so replace it with snprintf() and
eliminate an error condition to check.  Add a check for the
dag_rx_get_stream_count() invocation.  Make sure to call dag_close() and
fclose() even when failing with an error.

fixup error handling