DAG: Validate environment variables better. [skip ci]
Make the environment variables named constants.  Add a helper function
to convert strings to integers using a more thorough validation, such
that if it returns a non-negative integer for an input string, the
string definitely represents the number.  In dag_activate() reject
integers that are not one of the valid values.  In case of a validation
error include the rejected string value in the error message, not the
parsed integer value.

Before:
 # ERF_FCS_BITS=invalid tcpdump -ni dag0
(works as if ERF_FCS_BITS=0)
 # ERF_DONT_STRIP_FCS=never tcpdump -ni dag0
(works as if ERF_DONT_STRIP_FCS=1)

After:
 # ERF_FCS_BITS=invalid tcpdump -ni dag0
tcpdump: dag_activate dag0: invalid ERF_FCS_BITS value (invalid) in
  environment
 # ERF_DONT_STRIP_FCS=never tcpdump -ni dag0
tcpdump: dag_activate dag0: invalid ERF_DONT_STRIP_FCS value (never) in
  environment