CMake: fix to support CMake 3.31.
1) remove SOURCE parameter from add_custom_command().

It was never a valid parameter, but, prior to CMake 3.31, it was
silently ignored.  In 3.31 and later, it's treated as an error by
default.

See
https://github.com/the-tcpdump-group/libpcap/pull/1380#issuecomment-2471731576
for a case where this breaks something.

2) expand the "still support old versions of CMake, for the benefit of
people with older versions, but avoid deprecation warnings/errors"
section to handle the deprecation of pre-3.10 versions by 3.31.

(cherry picked from commit d920dd9945e60b9941cab6f0a8811a6c617fe592)