DAG: Fix ps_drop for stream drop counters. [skip ci]
The counter is supposed to mean the number of dropped packets since the
device activation.  When stream drop counters are available, this does
not hold because dag_activate() divides the stream number by two and
because dag_stats() copies the current value of the counter (which
persists between capture sessions), not the increment since the
activation, into the output.  Get it right and make the intermediate
counter values 64-bit.