CI: Remove VS 2015 build, add VS 2022 build.
VS 2015 is approximately 10 years old; get tid of the VS 2015 builds.

Add VS 2022 builds; try doing them with and without external capture
support.

Remove include of <wspiapi.h>; it's only needed for IPv6 support on
Windows 2000, which is almost 25 years old, and it has code that
provokes warnings in VS2022 builds.

In filtertest, don't arrange to call WSACleanup() via atexit(), as that
causes a pointer casting warning in VS 2022 builds.

(AppVeyor-only, so...)

[skip ci]