CMake: handle Windows architectures as we do for tcpdump.
Just use CMAKE_GENERATOR_PLATFORM to determine for which architecture
we're building, and use that to set the subdirectory of the SDK's Lib
directory in which to look for libraries.

Use the -A flag for all generators; according to the CMake
documentation, they're supported for all generators in CMake 3.1 and
later and, on Windows, we require CMake 3.12 or later.  That ensures
that CMAKE_GENERATOR_PLATFORM will be set, so that we can use it to
determine the right directory in which to look for pcap libraries.

(cherry picked from commit f7e5d8b991e2a619a37885185c620ee2f2395052)