pcap-filter(7): Clarify Ethernet, IPv4 and IPv6. [skip ci]
In "ether dst" remove the "various mixes of ':', '.', and '-'" because
that's not what the code implements.  In "ip proto" update the lists of
platforms that support one or another protocol name and describe the
dependency on getprotobyname(3) in more detail.  Document the "carp",
"vrrp", "igmp" and "igrp" abbreviations.  List two more OSes that
recognise "ipv6-icmp".