CI: reorganize AppVeyor tests.
Do the Visual Studio tests in 32-bit/64-bit pairs, with all other
settings the same.

For VS 2015 and VS 2017, do both WinPcap and Npcap, with AirPcap and
remote capture off.

For VS 2019, do both WinPcap and Npcap, with Npcap done both without
AirPcap and remote capture and with both AirPcap and remote capture.

(This is a subset of the full matrix, so that we don't take *too* much
time.)

Annotate each of the pairs.