Have pcapint_find_function() return a void * rather than a FARPROC.
Casting from FARPROC, which is the type of the return value of
GetProcAddress(), to a function pointer gets a C4191 warning
from Visual Studio 2022.

Casting FARPROC to void * and returning the result gets the
warning; we forcibly suppress it here.  This does mean that
the warnings aren't produced elsewhere, however, as casting
the void * to a function pointer doesn't get the same warning.

Given that, and given that the equivalent UN*X API, dlsym(),
returns a void *, we have pcapint_find_function() return
a void *.

(cherry picked from commit c1e73c9ba11718a4b77c9aa36a47869ed753267e)