Mind netmap support in pcap_lib_version().
It is not immediately obvious whether a particular build of libpcap is
expected to support netmap because netmap instead of introducing capture
devices allows to use existing network interfaces in netmap mode by
prepending "netmap:" to a supported interface name.  To clarify this, on
BPF platforms (i.e. FreeBSD) and Linux, if PCAP_SUPPORT_NETMAP is
defined, mention netmap in the result of pcap_lib_version() in addition
to any other features that may already appear there.

Before:
libpcap version 1.11.0-PRE-GIT (with zerocopy support)

After:
libpcap version 1.11.0-PRE-GIT (with zerocopy and netmap support)