Do not test for "dummy" in the interface name.
"dummy" is indeed the default prefix for Linux dummy network interfaces.
As far as libpcap is concerned, it is not different from other network
interfaces, so there is no good reason to skip it.  Fortunately, the
checks in fad-gifc.c and fad-glifc.c are dead code because Linux uses
fad-getad.c.