dlpi: fix open_dlpi_device() to handle Solaris/HP-UX logical interfaces.
If handed the name of a logical interface, strip off the logical
interface number.  This means that in pcap_findalldevs() we won't skip
the logical interfaces and will get the network addresses and netmasks
they have.

(It also mean that if you try to capture on a logical interface, it'll
capture on the physical interface, but that might be the right thing to
do.)

We can't do it earlier in fad-gifc.c and fad-glifc.c, because we have to
use the logical interface name when fetching the netmask, etc..

(cherry picked from commit 9b566c3239e34b2131fee9f2aebf7842b48bcdfe)