TESTrun: Implement expression aliases. [skip appveyor]
The filter language syntax often provides more than one way to spell the
same predicate whilst getting the same bytecode as the result.  In such
cases definining a separate unit test for every variant of the filter
expression is an unnecessary duplication and does not in fact verify
that the bytecode is identical.

To address this, introduce an optional "aliases" key in the unit test
hash, convert various existing aliases to the new encoding and add many
new aliases.  Split a few compound filter expressions and test the
elements one by one.  Rewrite some index expressions to cover the syntax
better.  Update various comments.