Support both "igrp" and "eigrp" in filters.
Use our own header to define IPPROTO_ values, rather than leaving them
up to the OS for which we're building.

Always use IP protocol 9 for "igrp", and use IP protocol 88 for "eigrp".
9 is assigned to "any private Internet gateway protocol", and is mainly
used for Cisco's IGRP.  88 is for Cisco's EIGRP, which is not compatible
on-the-wire with IGRP.

FreeBSD, DragonFly BSD and macOS define IPPROTO_IGRP as 88, rather than
as 9, and their libpcaps use IP protocol 88 for "igrp".  Everybody else
uses IP protocol 9 for "igrp".