TESTrun: Condense the output format. [skip appveyor]
Spending a line per unit test and printing the errors inline does not
work well for what is now more than 900 tests: if a test has failed,
getting to the details of it requires a lot of scrolling, and if all
tests have passed, there is no point in overwhelming an interactive
session or an automatic build log with so much output.

Replace the long list with a dense PHPUnit-style map, after it provide
details of any skipped or failed tests, similar to tcpdump TESTrun.