TESTrun: Make assorted minor improvements.
Add a BSD-2-Clause licence boilerplate.

Add a Perl version requirement (this may be not exact: 5.26.1 works and
5.22.1 does not, so there may be another working version in between, but
that can be confirmed later as and if required).  This produces a
sensible error message instead of syntax errors on Solaris 10, Ubuntu
16.04 etc.

When running TESTrun on a slow host, Ctrl-C interrupts its child
processes, but not the script.  Use sigtrap to make it easier to stop.

Specify a template for File::Temp, this way it is easier to clean up if
something went wrong.