TESTrun: Use tcpdump-like defaults for files. [skip appveyor]
In tcpdump TESTrun uses ./config.h and ./tcpdump, which works for build
matrix purposes.  It does not work when the current directory is not
where the build output files go to, but in practice that does not seem
to be a problem.  Use a similar logic in libpcap TESTrun and remove
CONFIG_H from the build matrix scope.