pcap-filter(7): Drop 4P man section references. [skip ci]
When discussing the /etc/services file, instead of tcp(4P) and udp(4P)
refer to the man page about the file, which is services(5) in *BSD,
Linux, illumos and Solaris 11.4.  In older Solaris versions it is
services(4), which users of such versions hopefully know.

When discussing IPv4, do not refer to the "ip" man page, which describes
the application of Berkeley sockets API to IPv4 and which appears in
section 4P in illumos and Solaris 11.4 only (section 7P in older Solaris
versions, 4 in *BSD and 7 in Linux).  In this context if the reader
wants to know more about IPv4, it is either the packet header structure
or addressing/routing principles, which has been discussed at length in
various publications, which are typically not man pages.