Factor TESTlib.pm out. [skip appveyor]
Rewrite the six skip subroutines in a generic manner (to take an
argument).  Add new subroutines to produce a test result of the required
type instead of hard-coding the hash structure and the ASCII character
in the test runner functions.  Move that and other generic reusable code
to the new file.  This way the project-specific script uses the shared
code roughly as follows:

1. Call read_config_h() if shared skip functions are required.
2. Define the test runner subroutines.
3. Produce a list of ready-to-run tests.
4. Call init_tmpdir().
5. Call test_and_report().