TESTrun: Require Perl version properly. [skip ci]
As Francois-Xavier notes, the current requirement does not work anymore
and TESTrun fails with a syntax error about TESTst.pm on Solaris 10,
which has Perl 5.8.4.  This is because the BEGIN code block, which
requires one of the helper files, executes before any other code.  Move
the Perl version requirement there to fix it.