filtertest: Add "-l" flag to use Linux BPF extensions.
filtertest prints what pcap_compile() produces for a pcap_t from
pcap_open_dead(), which is not suitable for Linux BPF extensions, so in
theory the output should never include any absolute addresses formatted
as names in square brackets ([ifindex], [vlanp], etc.).  In practice it
sometimes does, which means such tests do not cover the intended part of
the problem space (bytecode without the extensions).  Another problem
is, bytecode with the extensions is a sound part of the problem space
too, so ideally the filter tests should completely and separately cover
each part.

To begin untangling this, add a Linux-specfic flag to filtertest to make
it clear whether the bytecode is meant to use the extensions.  It has
the correct effect on the "vlan" keyword already and there are VLAN
tests without the extensions already, so add VLAN tests for bytecode
with the extensions and get both versions tested at basic level.