Simplify the resulting bytecode for "pppoes NUM".
Instead of loading a [32-bit] word and using only its low 16 bits load a
[16-bit] half-word with the same 16 bits and get an equivalent behaviour
without the bitwise AND instruction.  While at it, add a comment with
the packet header diagram and use a standard named constant for the
maximum session ID value.  Update bytecode in the associated test.

Using this revision of libpcap, tcpdump matches "pppoes 0x3b" in one of
its test files as before and passes its "make check".