Fix, test and document ARCnet address parsing.
ARCnet address parsing works incorrectly:

$ filtertest ARCNET 'link src $42'
(000) ldb      [0]
(001) jeq      #0xd4            jt 2	jf 3
(002) ret      #262144
(003) ret      #0

Before commit bcbef22 in October 2018 the lexer used pcap_ether_aton()
directly and in the case of an AID token pointed the function argument
one character past the beginning of the string to skip the initial "$".
In the example above the function would parse "42" into the first byte
of the MAC address (0x42) and return.  Then the lexer passed the token
with a partially initialized MAC address to the parser, which passed the
MAC address to gen_acode(), which passed it to gen_ahostop(), which
correctly used the initialized byte only.

Since the commit gen_acode() gets the original string and passes it to
pcap_ether_aton(), which in the example above initializes the first MAC
address nibble to 0xd, the second to 0x4 and the third (not shown) to
0x2.

Refactor this code to fix the bug and to fail more reliably.  Implement
a new pcapint_atoan() function to parse ARCnet addresses only and to
indicate whether the input string is valid.  In gen_acode() use the new
function instead of pcap_ether_aton() and check the return value before
using the result.  Ibid., simplify memory management by using a local
variable instead of the compiler state for the temporary parsed address.
In gen_ahostop() change the address size to 8 bits to have the solution
space match the problem space better.  Now it works as expected:

$ filtertest ARCNET 'link src $42'
(000) ldb      [0]
(001) jeq      #0x42            jt 2	jf 3
(002) ret      #262144
(003) ret      #0

Add accept/apply/reject filter tests to cover the updated implementation
and document ARCnet syntax in pcap-filter(7).  The
bacnet-arcnet-linux.pcap file is from the Wireshark packet capture
collection.