Use 32-bit comparisons for "decnet dst" when possible.
gen_dnhostop() already compares PLENGTH and FLAGS in parallel, so
compare the destination address as well when it is within reach (that
is, for the short frame encoding).  Fix and expand my earlier comment to
document gen_dnhostop() better.

This changes the resulting bytecode: it now takes 4 fewer instructions
to test a destination DECnet address (thus the corresponding accept
tests change), but the behaviour remains the same (thus no apply tests
change).