TESTrun: Accommodate older Perl versions.
Simplify Perl syntax in TESTst.pm such that Perl 5.8.4 can run it too;
this changes the test runner function argument data type a bit, so make
TESTmt.pm and TESTrun consistent in this regard.  In TESTmt.pm use a
token end-of-queue data item instead of Thread::Queue->end(), which is
not in Perl 5.10.1.  Join a tester thread right after receiving its
end-of-queue token.  Use old style for File::Temp and a couple of for
loops.  Switch from modern heredoc to plain multiline strings and a
converter foreach loop.  Use a workaround to count the number of items
in a hash.  With these changes TESTrun works as expected using:

* Perl 5.8.4 w/o threads (Solaris 10 default)
* Perl 5.10.1 w/threads (Solaris 10 with OpenCSW package)
* Perl 5.10.1 w/threads (Debian 6)