Make extended Linux VLAN tests conditional.
Old Linux kernels do not include VLAN information in the auxiliary data,
in which case libpcap does not use BPF extensions for VLAN filtering, in
which case TESTrun accept blocks for VLAN with BPF extensions try to
test libpcap code that isn't compiled in the first place.

In Autoconf and CMake check whether SKF_AD_VLAN_TAG_PRESENT is defined.
In TESTrun add a new have_decl() subroutine to match the semantics of
AC_CHECK_DECLS() and use it to skip the test blocks if appropriate.