Autoconf: Simplify the deciding of valgrindtest.
Knowing the OS is sufficient to tell whether to build valgrindtest, so
take VALGRINDTEST_SRC out of the packet capture type detection and make
it a single case statement just before the AC_SUBST().  This way the
code does the same, but is much easier to relate with the CMake leg.