Simplify the use of struct sockaddr_ll on Linux.
To work with AF_PACKET addresses, libpcap needs a declaration of struct
sockaddr_ll.  It originally appeared in Linux kernel 2.1.67 (1997-11-29)
in <linux/if_packet.h> and subsequently in GNU libc 2.0.93 (1998-03-26)
in <netpacket/packet.h>.  musl libc reproduced the latter in version
0.7.11 (2011-06-28) whilst still in its alpha series.  This way, now on
all supported Linux systems the build checks for <netpacket/packet.h>
are always true, so remove them from Autoconf and CMake.

Similarly, on all supported Linux systems AF_PACKET is always defined,
so do not check whether AF_PACKET is defined if __linux__ is defined.

Also Linux is not the only OS that defines AF_PACKET, but the only one
that uses it with struct sockaddr_ll, so guard the use of that structure
with "#ifdef __linux__" to clarify the scope.