Optimize packet capture type auto-detection.
When auto-detecting the packet capture type, Haiku always means using
pcap-haiku.c and vice versa, so instead of checking for a Haiku-specific
header on every OS (and not using the check result afterwards even on
Haiku) use the OS name, which the build system has already detected.
Apply the same logic to Linux and GNU/Hurd auto-detection.

Try auto-detecting these three OSes first since they don't require the
header checks now.  Then, if necessary, try BPF and check for
<net/bpf.h> and <sys/ioccom.h> only.  Then, if necessary, try DLPI and
check for <sys/dlpi.h> only.  This way the auto-detection result stays
the same as before, but the header checks occur only if required for the
build.