Optimize detection of some optional capture modules.
Check for RDMA, DAG and SNF on Linux only.  If the current OS is not
Linux, in the Autoconf leg change the default to "do not auto-detect"
and fail if the user has explicitly requested to enable the module, in
the CMake leg disable the checks and do not provide an option to enable
the module.  Do the same for netmap not on {FreeBSD, Linux, Windows}, as
well as for DPDK not on {FreeBSD, Linux}.