Simplify the use of getpass() on Haiku.
getpass() is in libbsd iff the OS is Haiku, so the check for getpass()
in libc is always false on Haiku and always true elsewhere, and the
check for getpass() in libbsd is always true on Haiku.  So, iff the OS
is Haiku, add libbsd to the list of required libraries and that's it.