Refine reporting of DPDK-related build messages.
As far as the two current build systems go, most optional modules print
a one-line message to stdout if the user did not specifically request
enabling the module and the module could not be detected automatically.
DPDK in this case prints a multi-line message to stderr, which draws
attention from more severe problems.

Make this behaviour consistent with other modules, also make sure that
in all cases the output indicates whether pkg-config was used, and what
the final outcome of the process is.  While at it, remove the QUIET
option from pkg-config detection to make it easier to understand why
pkg-config was used (or not).