Haiku: Look for ethers(5) in /boot/system/settings/network/.
Haiku does not have the file because it does not have ether_hostton(),
but if it had the file, it likely would be in that directory next to
hosts(5) and services(5) rather than in /etc.  Since in this case
libpcap defaults to pcap_ether_hostton(), which tries to read the
entries from somewhere anyway, have this [libpcap-only] somewhere fit
Haiku filesystem layout.