TESTrun: Use /etc/hosts and/or /etc/ethers if possible.
The tests currently done in TESTrun exercise the BPF compiler code paths
that are effectively pure functions, that is, do not depend on any
external effects.  This among other things avoids name-to-address
resolution because a random host that is running the tests cannot be
reasonably expected to resolve any hostnames reliably or at all.

However, name-to-address resolution is a common use case.  To make it
possible to test it too, introduce a set of test-only hostnames and
functions to detect whether these hostnames seem to be configured and
whether the host uses case-sensitive string comparison on Ethernet or
Internet(!) hostnames.  In the latter case, to enable skipping of tests
on musl libc, in Autoconf and CMake map __GLIBC__ and __UCLIBC__ from
<features.h> to HAVE_GLIBC and HAVE_UCLIBC respectively in config.h,
which is also available in the Perl solution space.

Add filter tests for variations of "ether host HOSTNAME", "ip host
HOSTNAME", "ip6 host HOSTNAME" and "gateway HOSTNAME".